add ability to define several server directives.
In case of using ELB, if you want to add redirect to url with www, normal rewrite rule will not work because ELB expects only 200 response code, but in case of rewrite rule nginx response with code 300. The solution is add additional server directive. I extend  template with variable server_list and add for loop.